consolidate legacy graphics compatibility
